server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi>=0.135.2",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.22",
    "langchain-google-genai>=4.2.1",    # embedding
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.1.12",
    "langchain-text-splitters>=1.1.1",
    "pinecone>=8.1.0",
    "pydantic>=2.12.5",                # typing
    "pymongo[bson]>=4.16.0",
    "pypdf>=6.9.2",                    #pdf parcing
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "requests>=2.33.0",
    "slowapi>=0.1.9",
    "tqdm>=4.67.3",                    # authentication
    "uvicorn>=0.42.0",
]



[tool.ruff.lint]
unfixable = ["F401"]
